+++ runtime/calamari/Cargo.toml
@@ -148,6 +148,7 @@ orml-xtokens = { git = "https://github.com/manta-network/open-runtime-module-lib
 
 # Self dependencies
 pallet-asset-manager = { path = '../../pallets/asset-manager' }
+runtime-common = { path = '../common' }
 
 
 [features]
